# fwber-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FWBer API Documentation",
        "description": "Comprehensive API documentation for FWBer - A proximity-based dating platform with real-time features, AI-powered matching, and privacy-first design.",
        "contact": {
            "name": "FWBer Team",
            "email": "support@fwber.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.fwber.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a message",
                "description": "Send a text message or media (image/video/audio/file) to a matched user. Either content or media is required. Enforces relationship tier limits and blocks.",
                "operationId": "1aa21572307a1c2e85281f5e04599031",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "receiver_id"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "description": "ID of message recipient",
                                        "type": "integer",
                                        "example": 42
                                    },
                                    "content": {
                                        "description": "Text message content",
                                        "type": "string",
                                        "maxLength": 5000,
                                        "example": "Hey, how are you?"
                                    },
                                    "message_type": {
                                        "type": "string",
                                        "enum": [
                                            "text",
                                            "image",
                                            "video",
                                            "audio",
                                            "file"
                                        ],
                                        "example": "text"
                                    },
                                    "media": {
                                        "description": "Media file (image/video/audio/file)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "media_duration": {
                                        "description": "Duration in seconds for audio/video",
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "sender_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "receiver_id": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Hey, how are you?"
                                        },
                                        "message_type": {
                                            "type": "string",
                                            "example": "text"
                                        },
                                        "media_url": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Messaging blocked or not allowed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Messaging blocked between users"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active match found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No active match found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/{userId}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get conversation with a matched user",
                "description": "Returns paginated messages and other user's presence info.",
                "operationId": "45f1f9530f5fbeccfeefc179cf845804",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        },
                                        "other_user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "last_seen_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Conversation access blocked"
                    },
                    "404": {
                        "description": "No active match"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/{messageId}/read": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Mark a message as read",
                "description": "Mark message as read",
                "operationId": "d04c1dac090d5ea63256813cdb3073f6",
                "parameters": [
                    {
                        "name": "messageId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked read"
                    },
                    "403": {
                        "description": "Only receiver can mark"
                    },
                    "404": {
                        "description": "Message not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/unread-count": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get count of unread messages",
                "description": "Get unread message count",
                "operationId": "bdd00f4d5bb74b58daa312d144a225ae",
                "responses": {
                    "200": {
                        "description": "Unread count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "unread_count": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UnauthorizedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}